refactor: 상태 관리하는 부분 타입스크립트로 변환
