+++ src/store/page.ts
@@ -1,0 +1,15 @@
+class Page {
+  #page: number;
+
+  constructor() {
+    this.#page = 1;
+  }
+
+  getPage() {
+    this.#page++;
+    return this.#page;
+  }
+}
+
+const page = new Page();
+export default page;
