+++ src/types/MovieType.ts
@@ -1,0 +1,7 @@
+interface MovieType {
+  title: string;
+  poster_path: string;
+  vote_average: number;
+}
+
+export default MovieType;
