--- src/utils/imageUrl.js
@@ -1,3 +1,0 @@
-const imageUrl = (path, size = 400) => `https://image.tmdb.org/t/p/w${size}${path}`;
-
-export default imageUrl;
