+++ src/utils/imageUrl.ts
@@ -1,0 +1,4 @@
+const imageUrl = (path: string, size: number = 400) =>
+  `https://image.tmdb.org/t/p/w${size}${path}`;
+
+export default imageUrl;
